refactor MakeRequest
