+++ pkg/analytics/checkandincrement/adapter.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package experiment
+package checkandincrement
 
 import (
 	"fmt"
@@ -22,26 +22,28 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 
-	"github.com/iter8-tools/iter8-controller/pkg/analytics/checkandincrement"
 	iter8v1alpha1 "github.com/iter8-tools/iter8-controller/pkg/apis/iter8/v1alpha1"
 )
 
 const (
+	// MetricsConfigMap ...
 	MetricsConfigMap = "iter8-metrics"
-	Iter8Namespace   = "iter8"
+	// Iter8Namespace ...
+	Iter8Namespace = "iter8"
 )
 
-func MakeRequest(instance *iter8v1alpha1.Experiment, baseline, experiment interface{}) (*checkandincrement.Request, error) {
+// MakeRequest ...
+func MakeRequest(instance *iter8v1alpha1.Experiment, baseline, experiment interface{}) (*Request, error) {
 	spec := instance.Spec
 
-	criteria := make([]checkandincrement.SuccessCriterion, len(spec.Analysis.SuccessCriteria))
+	criteria := make([]SuccessCriterion, len(spec.Analysis.SuccessCriteria))
 	for i, criterion := range spec.Analysis.SuccessCriteria {
 		iter8metric, ok := instance.Metrics[criterion.MetricName]
 		if !ok {
 			// Metric template not found
 			return nil, fmt.Errorf("Metric %s Not Available", criterion.MetricName)
 		}
-		criteria[i] = checkandincrement.SuccessCriterion{
+		criteria[i] = SuccessCriterion{
 			MetricName:         criterion.MetricName,
 			Type:               criterion.ToleranceType,
 			Value:              criterion.Tolerance,
@@ -56,14 +58,14 @@ func MakeRequest(instance *iter8v1alpha1.Experiment, baseline, experiment interf
 	now := time.Now().Format(time.RFC3339)
 	destinationKey, namespaceKey, baseVal, experimentVal, baseNsVal, experimentNsVal := "", "", "", "", "", ""
 	switch instance.Spec.TargetService.APIVersion {
-	case KubernetesService:
+	case "v1":
 		destinationKey = "destination_workload"
 		namespaceKey = "destination_service_namespace"
 		baseVal = baseline.(*appsv1.Deployment).GetName()
 		experimentVal = experiment.(*appsv1.Deployment).GetName()
 		baseNsVal = baseline.(*appsv1.Deployment).GetNamespace()
 		experimentNsVal = experiment.(*appsv1.Deployment).GetNamespace()
-	case KnativeServiceV1Alpha1:
+	case "serving.knative.dev/v1alpha1":
 		destinationKey = "destination_service_name"
 		namespaceKey = "destination_service_namespace"
 		baseVal = baseline.(*corev1.Service).GetName()
@@ -74,9 +76,9 @@ func MakeRequest(instance *iter8v1alpha1.Experiment, baseline, experiment interf
 		return nil, fmt.Errorf("Unsupported API Version %s", instance.Spec.TargetService.APIVersion)
 	}
 
-	return &checkandincrement.Request{
+	return &Request{
 		Name: instance.Name,
-		Baseline: checkandincrement.Window{
+		Baseline: Window{
 			StartTime: instance.ObjectMeta.GetCreationTimestamp().Format(time.RFC3339),
 			EndTime:   now,
 			Tags: map[string]string{
@@ -84,7 +86,7 @@ func MakeRequest(instance *iter8v1alpha1.Experiment, baseline, experiment interf
 				namespaceKey:   baseNsVal,
 			},
 		},
-		Candidate: checkandincrement.Window{
+		Candidate: Window{
 			StartTime: instance.ObjectMeta.GetCreationTimestamp().Format(time.RFC3339),
 			EndTime:   now,
 			Tags: map[string]string{
@@ -92,7 +94,7 @@ func MakeRequest(instance *iter8v1alpha1.Experiment, baseline, experiment interf
 				namespaceKey:   experimentNsVal,
 			},
 		},
-		TrafficControl: checkandincrement.TrafficControl{
+		TrafficControl: TrafficControl{
 			MaxTrafficPercent: instance.Spec.TrafficControl.GetMaxTrafficPercentage(),
 			StepSize:          instance.Spec.TrafficControl.GetStepSize(),
 			SuccessCriteria:   criteria,
